Add files via upload
This is basically what i pictured  for the controls, basically a click/press move (i had to use two different tutorials to figure it out), and of course how it shoots. Im going to add the collision that makes the ship spin and travel a certain distance/speed  out of the users control depending on how nay times its been hit. Im thinking i could use a timer for this, maybe stopping user control for a bit, but enough for them to regain control after a half second or something.