Update Gamestate.js
Extended the follow distance of the ship to the mouse from 5 pixels to 50.
Moved the rotation code of the ship/weapon outside of the mouse distance if loop. 
This gives a bit of empty space between the mouse pointer and the ship, and allows the ship to be rotated with the mouse at any distance.